requirements.txt
google-adk
google-cloud-bigquery
google-cloud-translate
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
